Move Graph to module.
